Ajust the humerger button
