+++ src/components/protectedRoute/PrivateRoutes.jsx
@@ -1,3 +1,4 @@
+import React, { useState } from 'react';
 import { Outlet, Navigate } from 'react-router-dom';
 import { useMediaQuery } from '@mui/material';
 import PropTypes from 'prop-types';
@@ -8,6 +9,11 @@ import './PrivateRoutes.scss';
 
 const PrivateRoutes = ({ isAllowed, children, redirectPath }) => {
   const TOKEN = localStorage.getItem('JWT_TOKEN');
+  const [showLeftbar, setShowLeftbar] = useState(false);
+  const handleHamburgerClick = () => {
+    setShowLeftbar(!showLeftbar);
+  };
+
   const isMobile = useMediaQuery('(max-width: 768px)');
   if (!isAllowed || TOKEN === 'null' || !TOKEN) {
     return <Navigate to={redirectPath} replace />;
@@ -15,7 +21,28 @@ const PrivateRoutes = ({ isAllowed, children, redirectPath }) => {
   return (
     children || (
       <section className="page-container">
-        {!isMobile && <LeftBar />}
+        {isMobile ? (
+          <>
+            <div
+              className={`hamburger-button${showLeftbar ? ' open' : ''}`}
+              onClick={handleHamburgerClick}
+              onKeyDown={(e) => {
+                if (e.key === 'Enter' || e.key === ' ') {
+                  handleHamburgerClick();
+                }
+              }}
+              role="button"
+              tabIndex={0}
+            >
+              <span />
+              <span />
+              <span />
+            </div>
+            {showLeftbar && <LeftBar />}
+          </>
+        ) : (
+          <LeftBar />
+        )}
         <div className="home">
           <Outlet />
         </div>
